underwent a lot of further analysis of DID method
used bacondecomp to look at what groups would have the largest impact on a normal TWFE regression. it was almost all in the comparison between states that expanded in 2014 and states that never expanded, which makes a lot of sense.

additionally, used event_study() and plot_event_study() to create the figure that has six estimator plots, allows for some comparisons. as to why the trend before the expansion is so different in Gardner than the others, i don't really know