finished up most of the code that compares the adjacent states
created all the maps and plots, did a lot of work on color schemes and such to make sure the maps are readable and consistent across each other. just need to do the synthetic DID, but that requires that I figure out how to make the covariate table, and I also want to extend the DIDTable back to 2006 since I'm able to, which would make for stronger parallel trends. I think I'm gonna do the DIDTable first