modifying figures for the poster, started(ish) the poster
I figured out that I can export the usmaps through inkscape as a .emf, which almost totally works except the patterned diagonal line i made in inkscape for missing county data does not convert. so, instead, i exported a png of the map into paint.net in order to create a mask, created a diagonal pattern in paint.net, and then make a .png of diagonal lines in the shape of the US. in inkscape i made the missing county color completely transparent, and then I can put the diagonal lines .png behind the imported .emf so that the map is vectorized but the diagonal lines are raster. potentially not worth the effort, but at least on the exported .pdf you'll be able to zoom into the map without pixelation and the .pdf file won't be ridiculously huge because vectored files are tiny.