did the same process for the heart disease map, added missing key
i don't know why, but the process for creating the checkered US background didn't work for heart disease- it was somehow a different shape. however, the one i made for insurance was able to be scaled to the correct size and worked perfectly fine. weird

also just added the diagonal lines to the key manually in inkscape