creating new black and white plots
the journal is not in color, so I need to modify all of my plots. i still need to do a lot of them, especially a lot of maps, which is unfortunate, since I don't really know how to do that

so far, it's kinda difficult, and i needed to change computers because my hard drive crapped out. hopefully i can get back to my normal laptop soon, but for now this works.